Reorganized to make it a bit easier to use.
